firebaseの型定義ファイルを作成・分離
