+++ src/components/utils/libs/firebase/index.ts
@@ -1,0 +1,2 @@
+export * from "./event";
+export * from "./users";
