+++ src/global-states/atoms.ts
@@ -1,24 +1,5 @@
 import { atom, useRecoilState } from "recoil";
-
-export type CurrentUser = {
-  active: boolean | undefined;
-  bio: string;
-  createdAt: Date;
-  displayName: string | undefined;
-  email: string | undefined;
-  faculty: string | null;
-  field: string | null;
-  fieldDetails: Array<string> | undefined;
-  github: string | undefined;
-  grade: string | null;
-  instagram: string | undefined;
-  photoURL: string;
-  position: number;
-  status: number;
-  twitter: string | undefined;
-  uid: string;
-  university: string;
-};
+import { CurrentUser } from "src/components/utils/libs/firebase/users";
 
 export const currentUserState = atom<CurrentUser | null>({
   key: "currentUserState",
